+++ crates/cli/tests/run_node_test.rs
@@ -1321,15 +1321,29 @@ mod affected_files {
             .assert();
         let output = get_assert_output(&assert);
 
-        // Order is not deterministic
-        assert!(predicate::str::contains("Args:").eval(&output));
-
         if cfg!(windows) {
-            assert!(predicate::str::contains(".\\input1.js").eval(&output));
-            assert!(predicate::str::contains(".\\input2.js").eval(&output));
+            assert!(predicate::str::contains("Args: .\\input1.js .\\input2.js").eval(&output));
         } else {
-            assert!(predicate::str::contains("./input1.js").eval(&output));
-            assert!(predicate::str::contains("./input2.js").eval(&output));
+            assert!(predicate::str::contains("Args: ./input1.js ./input2.js").eval(&output));
         }
+    }
+
+    #[test]
+    fn sets_env_var() {
+        let fixture = create_sandbox_with_git("node");
+
+        fs::write(fixture.path().join("base/input1.js"), "").unwrap();
+        fs::write(fixture.path().join("base/input2.js"), "").unwrap();
+
+        let assert = create_moon_command(fixture.path())
+            .arg("run")
+            .arg("node:affectedFilesEnvVar")
+            .arg("--affected")
+            .assert();
+        let output = get_assert_output(&assert);
+
+        assert!(
+            predicate::str::contains("MOON_AFFECTED_FILES=./input1.js,./input2.js").eval(&output)
+        );
     }
 }